requirements.txt
apache-airflow==2.7.2
apache-airflow-providers-google==10.11.1
apache-airflow-providers-snowflake==4.2.0
apache-airflow-providers-mysql==3.6.0
apache-airflow-providers-postgres==5.6.0
pandas==2.1.4
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
snowflake-connector-python==3.7.0
google-cloud-storage==2.14.0
